requirements.txt
discord.py>=0.16.7
read-env>=1.1.0
SQLAlchemy>=1.1.9
